restructuring components and pages
